requirements.txt
Flask==3.1.0
folium==0.18.0
google-generativeai==0.8.3
httpx==0.27.2
numpy==2.1.3
pandas==2.2.3
plotly==5.23.0
protobuf==5.28.3
streamlit==1.36.0
streamlit_folium==0.23.2
utils==1.0.2
tenacity==8.5.0
